ci: add CHROMA_URL to turbo
