+++ apps/api/test/test.setup.ts
@@ -1,5 +1,3 @@
-import 'dotenv/config'
-
 import { vi } from 'vitest'
 
 vi.mock('sendgrid', () => ({
