+++ apps/cli/src/commands/serve/routers/finance.command.ts
