+++ apps/cli/src/commands/serve/trpc.ts
@@ -1,0 +1,4 @@
+import { initTRPC } from '@trpc/server'
+
+// Initialize tRPC
+export const trpc = initTRPC.create()
