+++ turbo.json
@@ -1,39 +1,51 @@
 {
-	"$schema": "https://turbo.build/schema.json",
-	"globalDependencies": [".env.local", "**/.env.*local"],
-	"globalEnv": [
-		"APP_URL",
-		"NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY",
-		"CLERK_SECRET_KEY",
-		"COOKIE_DOMAIN",
-		"COOKIE_NAME",
-		"COOKIE_SALT",
-		"COOKIE_SECRET",
-		"DATABASE_URL",
-		"GITHUB_TOKEN",
-		"GOOGLE_CLIENT_ID",
-		"GOOGLE_CLIENT_SECRET",
-		"GOOGLE_REDIRECT_URI",
-		"JWT_SECRET",
-		"OPENAI_API_KEY",
-		"OPENAI_ORG_ID",
-		"REDIS_URL",
-		"SENDGRID_API_KEY"
-	],
-	"tasks": {
-		"build": {
-			"dependsOn": ["^build"],
-			"outputs": [".next/**", "!.next/cache/**", "build"]
-		},
-		"dev": {
-			"cache": false,
-			"persistent": true
-		},
-		"lint": {},
-		"migrate": {},
-		"start": {
-			"dependsOn": ["^build"]
-		},
-		"test": {}
-	}
+  "$schema": "https://turbo.build/schema.json",
+  "globalDependencies": [
+    ".env.local",
+    "**/.env.*local"
+  ],
+  "globalEnv": [
+    "APP_URL",
+    "NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY",
+    "CHROMA_URL",
+    "CLERK_SECRET_KEY",
+    "COOKIE_DOMAIN",
+    "COOKIE_NAME",
+    "COOKIE_SALT",
+    "COOKIE_SECRET",
+    "DATABASE_URL",
+    "GITHUB_TOKEN",
+    "GOOGLE_CLIENT_ID",
+    "GOOGLE_CLIENT_SECRET",
+    "GOOGLE_REDIRECT_URI",
+    "JWT_SECRET",
+    "OPENAI_API_KEY",
+    "OPENAI_ORG_ID",
+    "REDIS_URL",
+    "SENDGRID_API_KEY"
+  ],
+  "tasks": {
+    "build": {
+      "dependsOn": [
+        "^build"
+      ],
+      "outputs": [
+        ".next/**",
+        "!.next/cache/**",
+        "build"
+      ]
+    },
+    "dev": {
+      "cache": false,
+      "persistent": true
+    },
+    "lint": {},
+    "migrate": {},
+    "start": {
+      "dependsOn": [
+        "^build"
+      ]
+    },
+    "test": {}
+  }
 }
